+++ fixtures/coverall/tests/test_generated_bindings.rs
@@ -3,6 +3,7 @@ uniffi_macros::build_foreign_language_testcases!(
     [
         "tests/bindings/test_coverall.py",
         "tests/bindings/test_coverall.kts",
-        "tests/bindings/test_coverall.rb"
+        "tests/bindings/test_coverall.rb",
+        "tests/bindings/test_handlerace.kts",
     ]
 );