+++ uniffi_bindgen/src/bindings/kotlin/templates/wrapper.kt
@@ -24,6 +24,7 @@ import com.sun.jna.Structure
 import java.nio.ByteBuffer
 import java.nio.ByteOrder
 import java.util.concurrent.atomic.AtomicLong
+import java.util.concurrent.atomic.AtomicBoolean
 import java.util.concurrent.atomic.AtomicReference
 import java.util.concurrent.locks.ReentrantLock
 import kotlin.concurrent.withLock